refactor: isolate proposals list listeners
